refactor: extract tracking logic to useThankYouPage hook
- Move business logic from component to custom hook
- Improve separation of concerns and code reusability
- Component now focuses only on UI rendering
